(chore #9) Refactor Activities -> Activity
