+++ lib/constants/carbonmark.constants.ts
@@ -1,0 +1,8 @@
+export const ActivityActions = [
+  "UpdatedQuantity",
+  "UpdatedPrice",
+  "CreatedListing",
+  "DeletedListing",
+  "Purchase",
+  "Sold",
+] as const;
