+++ lib/types/carbonmark.ts
@@ -1,4 +1,5 @@
 import { BigNumber } from "ethers";
+import { ActivityActions } from "../constants/carbonmark.constants";
 
 export interface Project {
   id: string;
@@ -13,7 +14,7 @@ export interface Project {
   listings: Listing[];
   price: BigNumber;
   country: Country;
-  activities: Activity[];
+  activities: ActivityType[];
   updatedAt: string; // timestamp
 }
 
@@ -23,7 +24,7 @@ export interface User {
   description: string;
   wallet: string;
   listings: Listing[];
-  activities: Activity[];
+  activities: ActivityType[];
   assets: string[]; // token addresses
 }
 
@@ -59,9 +60,12 @@ export type Listing = {
   };
 };
 
-export type Activity = {
+/** A type representing possible activity actions (e.g "Sold", "Purchase" etc) */
+export type ActivityActionT = typeof ActivityActions[number];
+
+export type ActivityType = {
   id: number;
-  activityType: string; // CreatedListing, DeletedListing, UpdatedPrice ...
+  activityType: ActivityActionT; // CreatedListing, DeletedListing, UpdatedPrice ...
   amount: BigNumber | null;
   previousAmount: BigNumber | null;
   price: BigNumber | null;
